feat router 처리 및  수정, view 이름 수정
