+++ trippy-client/src/views/group-account/AgreementStep5View.vue
@@ -1,0 +1,42 @@
+<script setup>
+import { ref } from "vue";
+import accountList from "@/_dummy/accountList_dummy.json";
+import RepresentativeAccountList from "@/components/group-account/RepresentativeAccountList.vue";
+import { useGroupAccountStore } from "@/stores/groupAccountStore";
+import NextButton from "@/components/common/NextButton.vue";
+import router from "@/router";
+
+const groupAccountStore = useGroupAccountStore();
+
+const selectAccountNumber = ref("");
+const selectAccountBank = ref("");
+
+const selectAccount = (account) => {
+  selectAccountBank.value = account.bankName;
+  selectAccountNumber.value = account.account;
+  groupAccountStore.setRepresentativeAccount(selectAccountNumber.value, selectAccountBank.value);
+};
+</script>
+
+<template>
+  <div class="flex flex-col h-full w-full bg-white justify-between">
+    <div>
+      <div>
+        <p class="title1 text-center mt-40">대표계좌를 설정해 주세요</p>
+      </div>
+      <RepresentativeAccountList
+        :accountList="accountList"
+        @selectAccount="selectAccount"
+        :accountBank="selectAccountBank"
+        :accountNumber="selectAccountNumber"
+      />
+    </div>
+    <NextButton
+      :title="'다음'"
+      :disabled="!selectAccountBank"
+      @click="router.push({ name: 'group-account-create-complete' })"
+    />
+  </div>
+</template>
+
+<style scoped></style>
